feat: add lossy optimization variants (BF16/FP16/INT8/INT4 weights and KV cache)
- Implement 09 lossy optimization variants for performance comparison
- Add BF16 weight optimization (01_weight_bf16)
- Add FP16/KV cache optimization (02_kv_cache_fp16)
- Add INT8 KV cache optimization (03_kv_cache_int8)
- Add INT8 weight optimization (04_weight_int8)
- Add INT4 weight optimization (05_weight_int4)
- Add full FP16 pipeline (06_full_fp16)
- Add full INT8 pipeline (07_full_int8)
- Add Flash to Linear attention conversion (08_flash_to_linear)
- Add Linear to Flash attention conversion (09_linear_to_flash)
- Update CMakeLists and kernel implementations
- Performance test scripts and documentation

+++ src/backend/cuda/include/linear_attention_cuda.hpp
@@ -3,6 +3,7 @@
 #include <cuda_runtime.h>
 #include <cublas_v2.h>
 #include <vector>
+#include <cuda_bf16.h>
 
 namespace qwen {
 namespace cuda {
@@ -56,12 +57,12 @@ class CudaLinearAttention {
     int value_dim_;
     int conv_kernel_;
 
-    float* d_in_proj_qkv_weight_;
-    float* d_in_proj_a_weight_;
-    float* d_in_proj_b_weight_;
-    float* d_in_proj_z_weight_;
-    float* d_conv1d_weight_;
-    float* d_out_proj_weight_;
+    __nv_bfloat16* d_in_proj_qkv_weight_;
+    __nv_bfloat16* d_in_proj_a_weight_;
+    __nv_bfloat16* d_in_proj_b_weight_;
+    __nv_bfloat16* d_in_proj_z_weight_;
+    __nv_bfloat16* d_conv1d_weight_;
+    __nv_bfloat16* d_out_proj_weight_;
     float* d_a_log_;
     float* d_dt_bias_;
     float* d_norm_weight_;
